update feature controllers

+++ public/utils/redis.js
@@ -12,9 +12,9 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.connectClient = connectClient;
 exports.getValue = getValue;
 exports.setValue = setValue;
+exports.updateValue = updateValue;
 exports.deleteValue = deleteValue;
 exports.flushAll = flushAll;
-exports.updateValue = updateValue;
 const redis_1 = require("redis");
 const constants_1 = require("./constants");
 const client = (0, redis_1.createClient)({
@@ -45,6 +45,20 @@ client.on("connect", () => {
 client.on("ready", () => {
     console.log("Redis client ready");
 });
+const getCacheKey = (cacheType, { id, value }) => {
+    if (!id && !value)
+        return null;
+    if (cacheType === 'feature:featureId')
+        return value ? `feature:${value.featureId}` : `feature:${id}`;
+    if (cacheType === 'features:projectId')
+        return value ? `features:projectId:${value.projectId}` : `feature:${id}`;
+    if (cacheType === 'project:projectId')
+        return value ? `project:${value.projectId}` : `project:${id}`;
+    if (cacheType === 'req:reqId')
+        return value ? `req:${value.reqId}` : `project:${id}`;
+    if (cacheType === 'team:members:teamId')
+        return value ? `team:members:${value.teamId}` : `team:members:${id}`;
+};
 function pullFromCache(key) {
     return __awaiter(this, void 0, void 0, function* () {
         if (!constants_1.REDIS)
@@ -60,47 +74,50 @@ function pullFromCache(key) {
         }
     });
 }
-function getValue(key) {
+function getValue(cacheType, id) {
     return __awaiter(this, void 0, void 0, function* () {
         if (!constants_1.REDIS)
             return;
-        try {
-            yield connectClient();
-            const res = yield pullFromCache(key);
-            return res;
-        }
-        catch (err) {
-            console.error('Error getting value:', err);
-            return undefined;
-        }
+        const key = getCacheKey(cacheType, { id });
+        if (!key)
+            return;
+        yield connectClient();
+        const res = yield pullFromCache(key);
+        return res;
     });
 }
-function setValue(key, value) {
+function setValue(cacheType, value) {
     return __awaiter(this, void 0, void 0, function* () {
         if (!constants_1.REDIS)
             return;
-        try {
-            yield connectClient();
-            const stringifiedValue = JSON.stringify(value);
-            console.log('value', stringifiedValue);
-            yield client.set(key, stringifiedValue);
-        }
-        catch (err) {
-            console.error("Error setting value:", err);
-        }
+        const key = getCacheKey(cacheType, { value });
+        if (!key)
+            return;
+        yield connectClient();
+        const stringifiedValue = JSON.stringify(value);
+        yield client.set(key, stringifiedValue);
     });
 }
-function deleteValue(key) {
+function updateValue(cacheType, value) {
     return __awaiter(this, void 0, void 0, function* () {
         if (!constants_1.REDIS)
             return;
-        try {
-            yield connectClient();
-            yield client.del(key);
-        }
-        catch (err) {
-            console.error("Error deleting value:", err);
-        }
+        const key = getCacheKey(cacheType, { value });
+        if (!key)
+            return;
+        yield connectClient();
+        yield setValue(cacheType, value);
+    });
+}
+function deleteValue(cacheType, id) {
+    return __awaiter(this, void 0, void 0, function* () {
+        if (!constants_1.REDIS)
+            return;
+        const key = getCacheKey(cacheType, { id });
+        if (!key)
+            return;
+        yield connectClient();
+        yield client.del(key);
     });
 }
 function flushAll() {
@@ -114,14 +131,6 @@ function flushAll() {
         catch (err) {
             console.error("Error flushing all:", err);
         }
-    });
-}
-function updateValue(key, newValue) {
-    return __awaiter(this, void 0, void 0, function* () {
-        if (!constants_1.REDIS)
-            return;
-        yield connectClient();
-        yield setValue(key, JSON.stringify(newValue));
     });
 }
 exports.default = client;
